Renamed class attribute signed_W to W_signed
signed_W was confusing as it could refer to the signed connectivity
matrix rather than to whether this connectivity matrix is signed.
Renaming it to W_signed makes this (slightly) clearer.

The change furthermore moves the list of neurotransmitters for which
the sign will be negative out of the class body, to make it more visible.